feat: 게시글 출력 수정

+++ src/main/java/kr/or/ddit/empt/cte/service/CareerTechnicalEducationVO.java
@@ -14,7 +14,7 @@ public class CareerTechnicalEducationVO {
 	private String jtName;
 	private int jtQuota;
 	private String jtSchool;
-	private int jtScore;
+	private double jtScore;
 	private Date jtStartDate;
 	private String jtTarget;
 	private String jtUrl;
